Migrated (once again) to react-google-maps-api.
Currently the representation of the markers on the map is fixed (is just
a Google Maps marker), but changing the underlying representation of the
Image/MarkerIcon will make it easy to modify the icon that is displayed
on the Map.

Also, need to style the zoom buttons on the map. Should be identical in
functionality to how it was in Phase 1.